# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Trading Social Media App",
        "version": "0.1.0"
    },
    "paths": {
...
        "/news-service/api/v1/create-news/": {
            "post": {
                "tags": [
                    "News Service"
                ],
                "summary": "Create News",
                "operationId": "create_news_news_service_api_v1_create_news__post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_news_news_service_api_v1_create_news__post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/news-service/api/v1/get-all-news/": {
            "get": {
                "tags": [
                    "News Service"
                ],
                "summary": "Get All News",
                "operationId": "get_all_news_news_service_api_v1_get_all_news__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/NewsResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get All News News Service Api V1 Get All News  Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/news-service/api/v1/get-news/{news_id}/": {
            "get": {
                "tags": [
                    "News Service"
                ],
                "summary": "Get News",
                "operationId": "get_news_news_service_api_v1_get_news__news_id___get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "news_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "News Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/news-service/api/v1/delete-news/{news_id}/": {
            "delete": {
                "tags": [
                    "News Service"
                ],
                "summary": "Delete News",
                "operationId": "delete_news_news_service_api_v1_delete_news__news_id___delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "news_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "News Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Body_create_news_news_service_api_v1_create_news__post": {
                "properties": {
                    "news_title": {
                        "type": "string",
                        "title": "News Title"
                    },
                    "news_picture": {
                        "type": "string",
                        "format": "binary",
                        "title": "News Picture"
                    }
                },
                "type": "object",
                "required": [
                    "news_title",
                    "news_picture"
                ],
                "title": "Body_create_news_news_service_api_v1_create_news__post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "NewsResponse": {
                "properties": {
                    "news_title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "News Title"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/SUser"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "picture_of_news": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Picture Of News"
                    },
                    "date_pub": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Date Pub"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "date_pub"
                ],
                "title": "NewsResponse"
            },
...
            "SUser": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token"
                    },
                    "picture_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Picture Url"
                    }
                },
                "type": "object",
                "title": "SUser"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "http://localhost:8000/auth-login/"
                    }
                }
            }
        }
    }
}